+++ apps/AppWithWearable/lib/pages/device/settings.dart
