+++ apps/AppWithWearable/lib/pages/device/widgets/info_button.dart
