+++ apps/AppWithWearable/lib/pages/device/widgets/transcript.dart
