+++ apps/AppWithWearable/lib/pages/home/page.dart
@@ -1,229 +1,223 @@
-import 'dart:async';
-
 import 'package:flutter/material.dart';
-import 'package:flutter_blue_plus/flutter_blue_plus.dart';
 import 'package:friend_private/backend/api_requests/cloud_storage.dart';
-import 'package:friend_private/utils/ble/communication.dart';
-import 'package:friend_private/utils/ble/connected.dart';
-import 'package:friend_private/utils/ble/scan.dart';
+import 'package:friend_private/backend/preferences.dart';
+import 'package:friend_private/backend/storage/memories.dart';
+import 'package:friend_private/flutter_flow/flutter_flow_theme.dart';
+import 'package:friend_private/pages/chat/page.dart';
+import 'package:friend_private/pages/device/page.dart';
+import 'package:friend_private/pages/device/settings.dart';
+import 'package:friend_private/pages/device/widgets/transcript.dart';
+import 'package:friend_private/pages/memories/page.dart';
 import 'package:friend_private/utils/notifications.dart';
-import 'package:friend_private/widgets/blur_bot_widget.dart';
-import 'package:friend_private/widgets/scanning_animation.dart';
-import 'package:friend_private/widgets/scanning_ui.dart';
-import 'package:google_fonts/google_fonts.dart';
-import '/backend/schema/structs/index.dart';
-import '/flutter_flow/flutter_flow_theme.dart';
-import 'widgets/transcript.dart';
 
-class HomePage extends StatefulWidget {
-  final Function refreshMemories;
+class HomePageWrapper extends StatefulWidget {
   final dynamic btDevice;
-  final GlobalKey<TranscriptWidgetState> transcriptChildWidgetKey;
 
-  const HomePage({
-    super.key,
-    required this.btDevice,
-    required this.refreshMemories,
-    required this.transcriptChildWidgetKey,
-  });
+  const HomePageWrapper({super.key, this.btDevice});
 
   @override
-  State<HomePage> createState() => _HomePageState();
+  State<HomePageWrapper> createState() => _HomePageWrapperState();
 }
 
-class _HomePageState extends State<HomePage> {
-  BTDeviceStruct? _device;
+class _HomePageWrapperState extends State<HomePageWrapper> {
+  GlobalKey<TranscriptWidgetState> transcriptChildWidgetKey = GlobalKey();
+  int _selectedIndex = 1;
+  List<Widget> screens = [Container(), const SizedBox(), const SizedBox()];
+  List<MemoryRecord> memories = [];
+  bool deepgramApiIsVisible = false;
+  bool openaiApiIsVisible = false;
+  final _deepgramApiKeyController = TextEditingController();
+  final _openaiApiKeyController = TextEditingController();
+  final _gcpCredentialsController = TextEditingController();
+  final _gcpBucketNameController = TextEditingController();
+  final _customWebsocketUrlController = TextEditingController();
+  bool _useFriendApiKeys = true;
+  String _selectedLanguage = 'en';
 
-  final scaffoldKey = GlobalKey<ScaffoldState>();
-  final unFocusNode = FocusNode();
-
-  StreamSubscription<BluetoothConnectionState>? connectionStateListener;
-  StreamSubscription? bleBatteryLevelListener;
-  int batteryLevel = -1;
+  _initiateMemories() async {
+    memories = await MemoryStorage.getAllMemories(filterOutUseless: true);
+    setState(() {});
+  }
 
   @override
   void initState() {
+    // _refreshMemories();
+    _initiateMemories();
+    WidgetsBinding.instance.addPostFrameCallback((_) async {
+      requestNotificationPermissions();
+    });
     super.initState();
-    if (widget.btDevice != null) {
-      _device = BTDeviceStruct.maybeFromMap(widget.btDevice);
-      _initiateConnectionListener();
-      _initiateBleBatteryListener();
-    } else {
-      scanAndConnectDevice().then((friendDevice) {
-        if (friendDevice != null) {
-          setState(() {
-            _device = friendDevice;
-          });
-          _initiateConnectionListener();
-          _initiateBleBatteryListener();
-        }
-      });
+  }
+
+  Future<void> _showSettingsBottomSheet() async {
+    // Load API keys from shared preferences
+    final prefs = SharedPreferencesUtil();
+    _deepgramApiKeyController.text = prefs.deepgramApiKey;
+    _openaiApiKeyController.text = prefs.openAIApiKey;
+    _gcpCredentialsController.text = prefs.gcpCredentials;
+    _gcpBucketNameController.text = prefs.gcpBucketName;
+    _customWebsocketUrlController.text = SharedPreferencesUtil().customWebsocketUrl;
+    _selectedLanguage = prefs.recordingsLanguage;
+    _useFriendApiKeys = prefs.useFriendApiKeys;
+
+    await showModalBottomSheet(
+      context: context,
+      isScrollControlled: true,
+      backgroundColor: Colors.black,
+      shape: const RoundedRectangleBorder(
+        borderRadius: BorderRadius.only(
+          topLeft: Radius.circular(20.0),
+          topRight: Radius.circular(20.0),
+        ),
+      ),
+      builder: (BuildContext context) {
+        return PopScope(
+            canPop: true,
+            child: StatefulBuilder(
+              builder: (context, StateSetter setModalState) {
+                return SettingsBottomSheet(
+                  deepgramApiKeyController: _deepgramApiKeyController,
+                  openaiApiKeyController: _openaiApiKeyController,
+                  deepgramApiIsVisible: deepgramApiIsVisible,
+                  openaiApiIsVisible: openaiApiIsVisible,
+                  gcpCredentialsController: _gcpCredentialsController,
+                  gcpBucketNameController: _gcpBucketNameController,
+                  customWebsocketUrlController: _customWebsocketUrlController,
+                  selectedLanguage: _selectedLanguage,
+                  onLanguageSelected: (String value) {
+                    setModalState(() {
+                      _selectedLanguage = value;
+                    });
+                  },
+                  useFriendAPIKeys: _useFriendApiKeys,
+                  onUseFriendAPIKeysChanged: (bool? value) {
+                    setModalState(() {
+                      _useFriendApiKeys = value ?? true;
+                    });
+                  },
+                  deepgramApiVisibilityCallback: () {
+                    setModalState(() {
+                      deepgramApiIsVisible = !deepgramApiIsVisible;
+                    });
+                  },
+                  openaiApiVisibilityCallback: () {
+                    setModalState(() {
+                      openaiApiIsVisible = !openaiApiIsVisible;
+                    });
+                  },
+                  saveSettings: _saveSettings,
+                );
+              },
+            ));
+      },
+    );
+  }
+
+  void _saveSettings() async {
+    final prefs = SharedPreferencesUtil();
+    prefs.openAIApiKey = _openaiApiKeyController.text.trim();
+    prefs.gcpCredentials = _gcpCredentialsController.text.trim();
+    prefs.gcpBucketName = _gcpBucketNameController.text.trim();
+
+    bool requiresReset = false;
+    if (_selectedLanguage != prefs.recordingsLanguage) {
+      prefs.recordingsLanguage = _selectedLanguage;
+      requiresReset = true;
     }
-    authenticateGCP();
+    if (_deepgramApiKeyController.text != prefs.deepgramApiKey) {
+      prefs.deepgramApiKey = _deepgramApiKeyController.text.trim();
+      requiresReset = true;
+    }
+    if (_customWebsocketUrlController.text != prefs.customWebsocketUrl) {
+      prefs.customWebsocketUrl = _customWebsocketUrlController.text.trim();
+      requiresReset = true;
+    }
+    if (_useFriendApiKeys != prefs.useFriendApiKeys) {
+      requiresReset = true;
+      prefs.useFriendApiKeys = _useFriendApiKeys;
+    }
+    if (requiresReset) transcriptChildWidgetKey.currentState?.resetState();
+
+    if (_gcpCredentialsController.text.isNotEmpty && _gcpBucketNameController.text.isNotEmpty) {
+      authenticateGCP();
+    }
   }
 
-  _initiateBleBatteryListener() async {
-    if (bleBatteryLevelListener != null) return;
-    bleBatteryLevelListener = await getBleBatteryLevelListener(_device!, onBatteryLevelChange: (int value) {
-      setState(() {
-        batteryLevel = value;
-      });
+  void _onItemTapped(int index) {
+    setState(() {
+      _selectedIndex = index;
     });
-  }
-
-  _initiateConnectionListener() async {
-    connectionStateListener = getConnectionStateListener(_device!.id, () {
-      // when bluetooth disconnected we don't want to reset the BLE connection as there's no point, no device connected
-      // we don't want either way to trigger the websocket closed event, because it's closed on purpose
-      // and we don't want to retry the websocket connection or something
-      widget.transcriptChildWidgetKey.currentState?.resetState(resetBLEConnection: false);
-      setState(() {
-        _device = null;
-      });
-      bleBatteryLevelListener?.cancel();
-      bleBatteryLevelListener = null;
-      // Sentry.captureMessage('Friend Device Disconnected', level: SentryLevel.warning);
-      createNotification(title: 'Friend Device Disconnected', body: 'Please reconnect to continue using your Friend.');
-      scanAndConnectDevice().then((friendDevice) {
-        if (friendDevice != null) {
-          setState(() {
-            _device = friendDevice;
-          });
-          clearNotification(1);
-          _initiateBleBatteryListener();
-        }
-      });
-    }, () {
-      widget.transcriptChildWidgetKey.currentState?.resetState(resetBLEConnection: true);
-    });
-  }
-
-  @override
-  void dispose() {
-    unFocusNode.dispose();
-    connectionStateListener?.cancel();
-    bleBatteryLevelListener?.cancel();
-    super.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
-    return GestureDetector(
-      onTap: () => unFocusNode.canRequestFocus
-          ? FocusScope.of(context).requestFocus(unFocusNode)
-          : FocusScope.of(context).unfocus(),
-      child: Scaffold(
-        key: scaffoldKey,
-        backgroundColor: FlutterFlowTheme.of(context).primary,
-        body: Stack(
+    return Scaffold(
+      body: Center(
+        child: IndexedStack(
+          index: _selectedIndex,
           children: [
-            const BlurBotWidget(),
-            ListView(children: [
-              ..._getConnectedDeviceWidgets(),
-              _device != null
-                  ? TranscriptWidget(
-                      btDevice: _device!,
-                      key: widget.transcriptChildWidgetKey,
-                      refreshMemories: widget.refreshMemories,
-                    )
-                  : const SizedBox.shrink(),
-            ]),
+            MemoriesPage(
+              memories: memories,
+              refreshMemories: _initiateMemories,
+            ),
+            HomePage(
+                btDevice: widget.btDevice,
+                refreshMemories: _initiateMemories,
+                transcriptChildWidgetKey: transcriptChildWidgetKey),
+            const ChatPage(),
           ],
         ),
+      ),
+      appBar: AppBar(
+        automaticallyImplyLeading: false,
+        backgroundColor: FlutterFlowTheme.of(context).primary,
+        title: Text(['Memories', 'Device', 'Chat'][_selectedIndex]),
+        elevation: 2.0,
+        centerTitle: true,
+        actions: [
+          IconButton(
+            icon: const Icon(
+              Icons.settings,
+              color: Colors.white,
+              size: 30,
+            ),
+            onPressed: _showSettingsBottomSheet,
+          )
+        ],
+      ),
+      bottomNavigationBar: BottomNavigationBar(
+        type: BottomNavigationBarType.fixed,
+        backgroundColor: FlutterFlowTheme.of(context).primary,
+        elevation: 0,
+        items: const <BottomNavigationBarItem>[
+          BottomNavigationBarItem(
+            icon: Icon(Icons.history),
+            label: 'Memories',
+          ),
+          BottomNavigationBarItem(
+            icon: Icon(Icons.bluetooth_connected),
+            label: 'Device',
+          ),
+          BottomNavigationBarItem(
+            icon: Icon(Icons.chat),
+            label: 'Chat',
+          ),
+        ],
+        currentIndex: _selectedIndex,
+        selectedItemColor: Colors.white,
+        unselectedItemColor: Colors.grey.shade700,
+        onTap: _onItemTapped,
       ),
     );
   }
 
-  _getConnectedDeviceWidgets() {
-    if (_device == null) {
-      return [
-        const SizedBox(height: 64),
-        const ScanningAnimation(),
-        const ScanningUI(
-          string1: 'Looking for Friend wearable',
-          string2: 'Locating your Friend device. Keep it near your phone for pairing',
-        ),
-      ];
-    }
-    return [
-      const SizedBox(height: 64),
-      const Center(
-          child: ScanningAnimation(
-        sizeMultiplier: 0.4,
-      )),
-      const SizedBox(height: 16),
-      Center(
-          child: Text(
-        'Connected Device',
-        style: FlutterFlowTheme.of(context).bodyMedium.override(
-              fontFamily: 'SF Pro Display',
-              color: Colors.white,
-              fontSize: 29.0,
-              letterSpacing: 0.0,
-              fontWeight: FontWeight.w700,
-              useGoogleFonts: GoogleFonts.asMap().containsKey('SF Pro Display'),
-              lineHeight: 1.2,
-            ),
-        textAlign: TextAlign.center,
-      )),
-      const SizedBox(height: 8),
-      Row(
-        crossAxisAlignment: CrossAxisAlignment.center,
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-          Text(
-            '${_device?.name ?? 'Friend'} ~ ${_device?.id.split('-').last.substring(0, 6)}',
-            style: const TextStyle(
-              color: Color.fromARGB(255, 255, 255, 255),
-              fontSize: 16.0,
-              fontWeight: FontWeight.w500,
-              height: 1.5,
-            ),
-            textAlign: TextAlign.center,
-          ),
-          batteryLevel == -1 ? const SizedBox.shrink() : const SizedBox(width: 16.0),
-          batteryLevel == -1
-              ? const SizedBox.shrink()
-              : Container(
-                  decoration: BoxDecoration(
-                    color: Colors.transparent,
-                    borderRadius: BorderRadius.circular(30),
-                    border: Border.all(
-                      color: Colors.white,
-                      width: 1,
-                    ),
-                  ),
-                  padding: const EdgeInsets.all(8.0),
-                  child: Row(
-                    mainAxisSize: MainAxisSize.min,
-                    children: [
-                      Text(
-                        '${batteryLevel.toString()}%',
-                        style: const TextStyle(
-                          color: Colors.white,
-                          fontSize: 12,
-                          fontWeight: FontWeight.bold,
-                        ),
-                      ),
-                      const SizedBox(width: 8.0),
-                      Container(
-                        width: 10,
-                        height: 10,
-                        decoration: BoxDecoration(
-                          color: batteryLevel > 75
-                              ? const Color.fromARGB(255, 0, 255, 8)
-                              : batteryLevel > 20
-                                  ? Colors.yellow.shade700
-                                  : Colors.red,
-                          shape: BoxShape.circle,
-                        ),
-                      ),
-                    ],
-                  ),
-                )
-        ],
-      ),
-      const SizedBox(height: 64),
-    ];
+  @override
+  void dispose() {
+    _deepgramApiKeyController.dispose();
+    _openaiApiKeyController.dispose();
+    _gcpCredentialsController.dispose();
+    _gcpBucketNameController.dispose();
+    _customWebsocketUrlController.dispose();
+    super.dispose();
   }
 }
