+++ apps/AppWithWearable/lib/pages/onboarding/welcome/page.dart
